requirements.txt
astroid==2.4.1
attrs==19.3.0
bcrypt==3.1.7
boltons==20.1.0
cffi==1.14.0
click==7.1.2
configparser==5.0.0
couchbase==2.5.12
flasgger==0.9.4
Flask==1.1.2
Flask-Cors==3.0.8
gunicorn==20.0.4
importlib-metadata==1.6.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
mypy-extensions==0.4.3
pycparser==2.20
PyJWT==1.7.1
pylint==2.5.2
pyrsistent==0.16.0
PyYAML==5.3.1
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.2
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
